Extract help center loader
