+++ client/layout/index.jsx
@@ -1,10 +1,7 @@
 import config from '@automattic/calypso-config';
-import { HelpCenter } from '@automattic/data-stores';
-import { useLocale } from '@automattic/i18n-utils';
 import { isWithinBreakpoint, subscribeIsWithinBreakpoint } from '@automattic/viewport';
 import { useBreakpoint } from '@automattic/viewport-react';
 import { useShouldShowCriticalAnnouncementsQuery } from '@automattic/whats-new';
-import { useDispatch } from '@wordpress/data';
 import clsx from 'clsx';
 import PropTypes from 'prop-types';
 import { Component, useCallback, useEffect, useState } from 'react';
@@ -17,6 +14,7 @@ import QuerySiteAdminColor from 'calypso/components/data/query-site-admin-color'
 import QuerySiteFeatures from 'calypso/components/data/query-site-features';
 import QuerySiteSelectedEditor from 'calypso/components/data/query-site-selected-editor';
 import QuerySites from 'calypso/components/data/query-sites';
+import HelpCenterLoader from 'calypso/components/help-center-loader';
 import JetpackCloudMasterbar from 'calypso/components/jetpack/masterbar';
 import { withCurrentRoute } from 'calypso/components/route';
 import SympathyDevWarning from 'calypso/components/sympathy-dev-warning';
@@ -27,16 +25,14 @@ import EmptyMasterbar from 'calypso/layout/masterbar/empty';
 import MasterbarLoggedIn from 'calypso/layout/masterbar/logged-in';
 import OfflineStatus from 'calypso/layout/offline-status';
 import isA8CForAgencies from 'calypso/lib/a8c-for-agencies/is-a8c-for-agencies';
-import { getGoogleMailServiceFamily } from 'calypso/lib/gsuite';
 import isJetpackCloud from 'calypso/lib/jetpack/is-jetpack-cloud';
 import { isWcMobileApp, isWpMobileApp } from 'calypso/lib/mobile-app';
-import { onboardingUrl } from 'calypso/lib/paths';
 import isReaderTagEmbedPage from 'calypso/lib/reader/is-reader-tag-embed-page';
 import { getMessagePathForJITM } from 'calypso/lib/route';
 import UserVerificationChecker from 'calypso/lib/user/verification-checker';
 import { useSelector } from 'calypso/state';
 import { isOffline } from 'calypso/state/application/selectors';
-import { isUserLoggedIn, getCurrentUser } from 'calypso/state/current-user/selectors';
+import { isUserLoggedIn } from 'calypso/state/current-user/selectors';
 import {
 	getShouldShowCollapsedGlobalSidebar,
 	getShouldShowGlobalSidebar,
@@ -47,7 +43,6 @@ import { getCurrentOAuth2Client } from 'calypso/state/oauth2-clients/ui/selector
 import getCurrentQueryArguments from 'calypso/state/selectors/get-current-query-arguments';
 import getIsBlazePro from 'calypso/state/selectors/get-is-blaze-pro';
 import getPrimarySiteSlug from 'calypso/state/selectors/get-primary-site-slug';
-import hasCancelableUserPurchases from 'calypso/state/selectors/has-cancelable-user-purchases';
 import isAtomicSite from 'calypso/state/selectors/is-site-automated-transfer';
 import isWooCommerceCoreProfilerFlow from 'calypso/state/selectors/is-woocommerce-core-profiler-flow';
 import { getIsOnboardingAffiliateFlow } from 'calypso/state/signup/flow/selectors';
@@ -76,8 +71,6 @@ import '@automattic/components/src/button/style.scss';
 import '@automattic/components/src/card/style.scss';
 
 import './style.scss';
-
-const HELP_CENTER_STORE = HelpCenter.register();
 
 function SidebarScrollSynchronizer() {
 	const isNarrow = useBreakpoint( '<660px' );
@@ -135,39 +128,17 @@ function WhatsNewLoader( { loadWhatsNew, siteId } ) {
 	);
 }
 
-function HelpCenterLoader( { sectionName, loadHelpCenter, currentRoute } ) {
-	const { setShowHelpCenter } = useDispatch( HELP_CENTER_STORE );
-	const isDesktop = useBreakpoint( '>782px' );
-	const handleClose = useCallback( () => {
-		setShowHelpCenter( false );
-	}, [ setShowHelpCenter ] );
-
-	const locale = useLocale();
-	const hasPurchases = useSelector( hasCancelableUserPurchases );
-	const user = useSelector( getCurrentUser );
+function HelpCenterLoaderWrapper( { sectionName, loadHelpCenter, currentRoute } ) {
 	const selectedSite = useSelector( getSelectedSite );
 	const primarySiteSlug = useSelector( getPrimarySiteSlug );
 	const primarySite = useSelector( ( state ) => getSiteBySlug( state, primarySiteSlug ) );
 
-	if ( ! loadHelpCenter ) {
-		return null;
-	}
-
 	return (
-		<AsyncLoad
-			require="@automattic/help-center"
-			placeholder={ null }
-			handleClose={ handleClose }
-			currentRoute={ currentRoute }
-			locale={ locale }
-			sectionName={ sectionName }
+		<HelpCenterLoader
 			site={ selectedSite || primarySite }
-			currentUser={ user }
-			hasPurchases={ hasPurchases }
-			// hide Calypso's version of the help-center on Desktop, because the Editor has its own help-center
-			hidden={ sectionName === 'gutenberg-editor' && isDesktop }
-			onboardingUrl={ onboardingUrl() }
-			googleMailServiceFamily={ getGoogleMailServiceFamily() }
+			sectionName={ sectionName }
+			loadHelpCenter={ loadHelpCenter }
+			currentRoute={ currentRoute }
 		/>
 	);
 }
@@ -340,7 +311,7 @@ class Layout extends Component {
 					loadWhatsNew={ loadHelpCenter && ! this.props.sidebarIsHidden && ! this.props.isNewUser }
 					siteId={ this.props.siteId }
 				/>
-				<HelpCenterLoader
+				<HelpCenterLoaderWrapper
 					sectionName={ this.props.sectionName }
 					loadHelpCenter={ loadHelpCenter }
 					currentRoute={ this.props.currentRoute }
